3.cpp fifth try

+++ 3/3.cpp
@@ -8,7 +8,7 @@ long long int e;
 int n, c, p;
 
 bool compare(int a, int b) {
-	long long int ans = (c * (a - b)) % p;
+	long long int ans = (c * (a - b)) % (long long int)p;
 	long long int power = e;
 	long long int base = a + b;
 	if (ans < 0)
@@ -16,14 +16,14 @@ bool compare(int a, int b) {
 	while (power > 0) {
 		if (power % 3 == 0) {
 			power /= 3;
-			base = (base * base * base) % p;
+			base = (base * base * base) % (long long int)p;
 		}
 		else if (power % 3 == 2){
-			ans = (ans * base * base) % p;
+			ans = (ans * base * base) % (long long int)p;
 			power -= 2;
 		}
 		else {
-			ans = (ans * base) % p;
+			ans = (ans * base) % (long long int)p;
 			power--;
 		}
 	}
